+++ test/fixtures/cookbooks/fake/metadata.rb
@@ -2,3 +2,4 @@ name 'fake'
 version '1.0.0'
 
 depends 'maven'
+depends 'java'